requirements.txt
altgraph==0.10.2
autopep8==1.3.1
bdist-mpkg==0.5.0
bonjour-py==0.3
boto==2.46.1
click==6.7
configparser==3.5.0
enum34==1.1.6
flake8==3.3.0
Flask==0.12.1
Flask-OAuth==0.12
google-api-python-client==1.6.2
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.6
macholib==1.5.1
MarkupSafe==1.0
matplotlib==1.3.1
mccabe==0.6.1
modulegraph==0.10.4
numpy==1.8.0rc1
oauth2==1.9.0.post1
oauth2client==4.1.0
py2app==0.7.3
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pyflakes==1.5.0
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
redis==2.10.5
rsa==3.4.2
scipy==0.13.0b1
six==1.10.0
uritemplate==3.0.0
virtualenv==15.1.0
Werkzeug==0.12.1
xattr==0.6.4
zope.interface==4.1.1
